+++ resources/views/components/livewire/attendance/poll-basics.blade.php
@@ -50,9 +50,10 @@
 
         <div class="mt-4">
             <x-input-label for="closing_at" :value="__('Closing at')"/>
-            <x-input type="datetime-local" id="closing_at" name="closing_at" class="mt-1 block w-full"
-                              wire:model="pollForm.closing_at"
-                              required autofocus autocomplete="closing_at"/>
+            <x-input-date-time id="closing_at"
+                               wire:model.blur="pollForm.closing_at"
+                               class="mt-1 block w-full"
+                               required/>
             @error('pollForm.closing_at')
             <x-input-error class="mt-2" :messages="$message"/>@enderror
         </div>